feat: Criando contexto de transação
