+++ 03-dt-money/src/App.tsx
@@ -3,13 +3,16 @@ import { GlobalStyle } from './styles/global'
 import { defaultTheme } from './styles/themes/default'
 import { Transactions } from './pages/Transactions'
 import { SummaryContainer } from './components/Summary/styles'
+import { TransactionsProvider } from './contexts/TransactionsContext'
 
 export function App() {
   return (
     <ThemeProvider theme={defaultTheme}>
       <GlobalStyle />
-      <SummaryContainer />
-      <Transactions />
+      <TransactionsProvider>
+        <SummaryContainer />
+        <Transactions />
+      </TransactionsProvider>
     </ThemeProvider>
   )
 }
