+++ 03-dt-money/src/contexts/TransactionsContext.tsx
@@ -1,0 +1,42 @@
+import { ReactNode, createContext, useEffect, useState } from 'react'
+
+interface Transaction {
+  id: number
+  description: string
+  type: 'income' | 'outcome'
+  category: string
+  price: number
+  createdAt: string
+}
+
+interface TransactionContextType {
+  transactions: Transaction[]
+}
+
+interface TransactionContextProviderProps {
+  children: ReactNode
+}
+
+export const TransactionContext = createContext({} as TransactionContextType)
+
+export function TransactionsProvider({
+  children,
+}: TransactionContextProviderProps) {
+  const [transactions, setTransactions] = useState<Transaction[]>([])
+  async function loadTransactions() {
+    const response = await fetch('http://localhost:3333/transactions')
+    const data = await response.json()
+
+    setTransactions(data)
+  }
+
+  useEffect(() => {
+    loadTransactions()
+  }, [])
+
+  return (
+    <TransactionContext.Provider value={{ transactions }}>
+      {children}
+    </TransactionContext.Provider>
+  )
+}
